Add ZSH config for setting environment variables on macOS Catalina and newer
On macOS Catalina and newer, Apple switched from Bash to ZSH. The new file for ".bash_profile" is ".zprofile".

Also the AnroidSDK path is different on a current macOS with the latest Android Studio, but I don't know since when this was changed, or if it was always like this.